--- src/pages/Auth/Login/LoginPage.styled.js
@@ -1,23 +1,0 @@
-import { styled } from 'styled-components';
-
-const LoginPageWrap = styled.div`
-  margin: 5.4rem 3.4rem;
-`;
-const PageH2 = styled.h2`
-  font-size: 2.4rem;
-  text-align: center;
-`;
-
-const TextInputBox = styled.div`
-  margin-top: 4rem;
-  margin-bottom: 3rem;
-`;
-
-const LoginCreateAccount = styled.a`
-  display: block;
-  margin-top: 20px;
-  color: #767676;
-  text-align: center;
-`;
-
-export { LoginPageWrap, PageH2, TextInputBox, LoginCreateAccount };
