+++ src/pages/Auth/ProfileSetting/ProfileSetting.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function ProfileSetting() {
+  return <div>초기 프로필 설정 페이지 입니다</div>;
+}
