+++ src/pages/Chat/ChatList/ChatList.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function ChatList() {
+  return <div>ChatList</div>;
+}
