+++ src/pages/Chat/ChatRoom/ChatRoom.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function ChatRoom() {
+  return <div>ChatRoom</div>;
+}
