--- src/pages/JoinMembers/JoinMembers.styled.js
@@ -1,22 +1,0 @@
-import { styled } from 'styled-components';
-
-const JoinMembersWrap = styled.div`
-  margin: 5.4rem 3.4rem;
-`;
-const PageH2 = styled.h2`
-  font-size: 2.4rem;
-  text-align: center;
-`;
-
-const TextInputBox = styled.div`
-  margin-top: 3rem;
-  margin-bottom: 3rem;
-`;
-
-const ErrorMessage = styled.p`
-  color: #eb5757;
-  font-size: 1.4rem;
-  margin-bottom: 2rem;
-`;
-
-export { JoinMembersWrap, PageH2, TextInputBox, ErrorMessage };
