+++ src/pages/Post/PostDetail/PostDetail.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function PostDetail() {
+  return <div>PostDetail</div>;
+}
