+++ src/pages/Post/PostEdit/PostEdit.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function PostEdit() {
+  return <div>PostEdit</div>;
+}
