+++ src/pages/Post/PostUpload/PostUpload.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function PostUpload() {
+  return <div>PostUpload</div>;
+}
