+++ src/pages/Product/ProductEdit/ProductEdit.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function ProductEdit() {
+  return <div>ProductEdit</div>;
+}
