+++ src/pages/Product/ProductUpload/ProductUpload.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function ProductUpload() {
+  return <div>ProductUpload</div>;
+}
