+++ src/pages/Profile/Follower/Follower.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Follower() {
+  return <div>Follower</div>;
+}
