+++ src/pages/Profile/ProfileDetail/ProfileDetail.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function MyProfile() {
+  return <div>MyProfile</div>;
+}
