+++ src/pages/Profile/ProfileEdit/ProfileEdit.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function ProfileEdit() {
+  return <div>ProfileEdit</div>;
+}
