+++ src/pages/Search/Search.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Search() {
+  return <div>Search</div>;
+}
