+++ src/pages/Splash/Splash.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Splash() {
+  return <div>Splash</div>;
+}
